add course mapper and make change in course dto and entity

+++ src/main/java/com/example/studentcourseregistration/dto/course/CourseResponse.java
@@ -14,9 +14,9 @@ public record CourseResponse(
 
         Integer creditHours,
 
-        Integer capacity,
+        Long capacity,
 
-        Integer remainingSeats,
+        Long remainingSeats,
 
         Term term,
